Separate FieldListItem to its own file, update names of columns and localize
