+++ ui/js/dfv/pods-dfv.min.asset.json
@@ -1,1 +1,1 @@
-{"dependencies":["wp-api-fetch","wp-autop","wp-components","wp-compose","wp-data","wp-element","wp-i18n"],"version":"8272b50d2afc2e142dee2dafd95f374a"}
+{"dependencies":["wp-api-fetch","wp-autop","wp-components","wp-compose","wp-data","wp-element","wp-i18n"],"version":"8cd40d72d208877aa0f82f49ac9f2e36"}
